back to windows and staying that way

+++ incl/KrimzLib/string.h
@@ -6,11 +6,11 @@ namespace kl {
 	public:
 		// Returns the file extension from the given file path
 		static std::string GetFileExtension(std::string filePath) {
-			int lastDotPos = filePath.find_last_of(".");
+			size_t lastDotPos = filePath.find_last_of(".");
 			if(lastDotPos == std::string::npos) {
 				return "";
 			}
-  			return filePath.substr(lastDotPos + 1);
+  			return filePath.substr(lastDotPos + 1L);
 		}
 	};
 }
